+++ ProyectoClipMoney2020/Models/Clases/Nacionalidad.cs
@@ -7,7 +7,7 @@ namespace ProyectoClipMoney2020.Models
 {
     public class Nacionalidad
     {
-        public int idNacionalidad { get; set; }
+        public long idNacionalidad { get; set; }
         public string descripcionNacionalidad { get; set; }
     }
 }
